Fixed a problem with the subdomain marking
THIS IS STILL NOT NICELY DONE
There appear to be a problem with marking the subdomains: some of the
vertices (the ones at the boundaries between two parts) are affected
random numbers.
A solution might be to change the "inside" methods by x + epsilon >=
something and x - epsilon <= something + step.