The Testing setup now includes the following three scenarios: - Original Operator in an UNWEIGHTED scenario - Original Operator in a WEIGHTED scenario - New alternativ Operator in a WEIGHTED Scenario
The two convex algorithms only work in the Original Scenario because they need access to a fully generated matrix for solving the convex problem.
The two convex algorithms also seem a lot slower than the other algorithms and the results do not share the accuracy of the other algorithms.

The are main reason for the failing test is that the assertion measures with an accuracy of 1e-10 yet some algorithm only approximate the ground truth up to
1e-6 or 1e-5. Therefore it might be worth to reduce the accuracy in the assertion point.

Also there seems to be a problem with WIHT in the unweighted case. It diverges every time.

The last thing is that we maybe should change the initialization of the parameter m in the unweighted case because using the function from WR produces an m>>n which is not
wanted.
Since the parameter m, n and s correlate we probably have to change s as well and maybe use the ground truth creation function with unweighted parameters.